requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
alembic==1.13.1
antlr4-python3-runtime==4.9.3
asteroid-filterbanks==0.4.0
async-timeout==4.0.3
attrs==23.2.0
audioread==3.0.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cmake==3.28.3
colorama==0.4.6
colorlog==6.8.2
contourpy==1.2.0
cycler==0.12.1
Cython==3.0.9
decorator==5.1.1
docopt==0.6.2
dtw-python==1.3.1
einops==0.7.0
filelock==3.13.1
fonttools==4.49.0
frozenlist==1.4.1
fsspec==2024.2.0
greenlet==3.0.3
huggingface-hub==0.21.3
HyperPyYAML==1.2.2
idna==3.6
importlib_resources==6.1.2
Jinja2==3.1.3
joblib==1.3.2
julius==0.2.7
kiwisolver==1.4.5
lazy_loader==0.3
librosa==0.10.1
lightning==2.2.0.post0
lightning-utilities==0.10.1
lit==17.0.6
llvmlite==0.42.0
Mako==1.3.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.3
mdurl==0.1.2
more-itertools==10.2.0
mpmath==1.3.0
msgpack==1.0.7
multidict==6.0.5
networkx==3.2.1
numba==0.59.0
numpy==1.26.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu11==10.2.10.91
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu11==11.7.4.91
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu11==2.14.3
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu11==11.7.91
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
openai-whisper==20231117
optuna==3.5.0
packaging==23.2
pandas==2.2.1
pillow==10.2.0
platformdirs==4.2.0
pooch==1.8.1
praat-parselmouth==0.4.3
primePy==1.3
protobuf==4.25.3
pyannote.audio==3.1.1
pyannote.core==5.0.0
pyannote.database==5.0.1
pyannote.metrics==3.2.1
pyannote.pipeline==3.0.1
pycparser==2.21
pydub==0.25.1
Pygments==2.17.2
pyparsing==3.1.1
python-dateutil==2.8.2
pytorch-lightning==2.2.0.post0
pytorch-metric-learning==2.4.1
pytz==2024.1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
rich==13.7.1
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
scikit-learn==1.4.1.post1
scipy==1.12.0
semver==3.0.2
sentencepiece==0.2.0
shellingham==1.5.4
six==1.16.0
sortedcontainers==2.4.0
soundfile==0.12.1
soxr==0.3.7
speechbrain==1.0.0
SQLAlchemy==2.0.27
sympy==1.12
tabulate==0.9.0
tensorboardX==2.6.2.2
threadpoolctl==3.3.0
tiktoken==0.6.0
torch==2.0.0+cu118
torch-audiomentations==0.11.1
torch-pitch-shift==1.2.4
torchaudio==2.0.1+cu118
torchmetrics==1.3.1
torchvision==0.15.1+cu118
tqdm==4.66.2
triton==2.0.0
typer==0.9.0
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
whisper-timestamped==1.15.1
yarl==1.9.4
zipp==3.17.0
